Backend/aiventory/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
protobuf==6.31.0
pymongo==4.13.0
python_bcrypt==0.3.2
django-cors-headers
google-auth
requests
traceback2